More implementation
Almost done with the server .. its much cleaner now

+++ include/INetworkHandlers.h
@@ -8,7 +8,7 @@ namespace SL {
 			virtual ~INetworkHandlers() {}
 
 			virtual void onConnection(const std::shared_ptr<ISocket>& socket) = 0;
-			virtual void onMessage(const std::shared_ptr<ISocket>& socket, const char* data, int len) = 0;
+			virtual void onMessage(const std::shared_ptr<ISocket>& socket, const char* data, size_t len) = 0;
 			virtual void onDisconnection(const ISocket* socket) = 0;
 		};
 	}
